SDK-921: Organise test directory
