+++ tests/data_type/image.spec.js
@@ -1,7 +1,7 @@
 const expect = require('chai').expect;
-const Image = require('../src/data_type/image');
-const ImageJpeg = require('../src/data_type/image.jpeg');
-const ImagePng = require('../src/data_type/image.png');
+const Image = require('../../src/data_type/image');
+const ImageJpeg = require('../../src/data_type/image.jpeg');
+const ImagePng = require('../../src/data_type/image.png');
 
 const imageContent = {
   toBase64() {
