+++ tests/request/index.spec.js
@@ -2,16 +2,16 @@ const expect = require('chai').expect;
 const nock = require('nock');
 const fs = require('fs');
 
-const config = require('../config');
-const request = require('../src/request');
+const config = require('../../config');
+const request = require('../../src/request');
 
-const AmlAddress = require('../src/aml_type').AmlAddress;
-const AmlProfile = require('../src/aml_type').AmlProfile;
-const Payload = require('../src/request/payload').Payload;
+const AmlAddress = require('../../src/aml_type').AmlAddress;
+const AmlProfile = require('../../src/aml_type').AmlProfile;
+const Payload = require('../../src/request/payload').Payload;
 
 const privateKeyFile = fs.readFileSync('./tests/sample-data/keys/node-sdk-test.pem', 'utf8');
 
-const yotiPackage = require('../package.json');
+const yotiPackage = require('../../package.json');
 
 describe('YotiResponse', () => {
   describe('#makeRequest', () => {
